Create search component
