+++ src/css/Home.css
@@ -1,7 +1,3 @@
-.search {
-  padding: 2em 0;
-}
-
 .content {
   margin: 0px auto;
   min-height: calc(100vh - 380px);
