+++ src/css/Searchbar.css
@@ -1,0 +1,3 @@
+.search {
+  padding: 2em 0;
+}
