+++ src/pages/Home.js
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
-import { Button, Dropdown, Input, Menu, Row } from "antd";
-import { DownOutlined } from "@ant-design/icons";
 
 import Footer from "../components/Footer";
+import Searchbar from "../components/Searchbar";
 import Header from "../components/Header";
 import Nav from "../components/Nav";
 import Synonyms from "../components/Synonyms";
@@ -10,39 +9,15 @@ import Translations from "../components/Translations";
 
 import "../css/Home.css";
 
-const { Search } = Input;
-
 function Home() {
   const [word, setWord] = useState("");
   const [find, setFind] = useState("Synonyms");
-
-  const onClick = ({ key }) => {
-    setFind(key);
-  };
-
-  const menu = (
-    <Menu onClick={onClick}>
-      <Menu.Item key="Synonyms">Synonyms</Menu.Item>
-      <Menu.Item key="Translations">Translations</Menu.Item>
-    </Menu>
-  );
 
   return (
     <>
       <Nav />
       {!word && <Header />}
-      <Row type="flex" align="center" className="search">
-        <Dropdown overlay={menu}>
-          <Button>
-            {find} <DownOutlined />
-          </Button>
-        </Dropdown>
-        <Search
-          placeholder="A word that describes your idea"
-          onSearch={(value) => setWord(value)}
-          style={{ width: 400 }}
-        />
-      </Row>
+      <Searchbar setWord={setWord} find={find} setFind={setFind} />
       <div className="content">
         {word && (
           <>
